Change byte array to int array in toBytes method for parity calculation

+++ src/main/java/uk/org/okapibarcode/backend/QrCode.java
@@ -1980,10 +1980,10 @@ public class QrCode extends Symbol {
             throw new OkapiInputException("Unable to determine ECI mode");
         }
 
-        byte[] bytes = toBytes(content, eci.charset);
+        int[] bytes = toBytes(content, eci.charset);
 
         int parity = 0;
-        for (byte b : bytes) {
+        for (int b : bytes) {
             parity ^= b & 0xFF;
         }
